Simplify d3d9cg.c/d3d9hlsl.c
