+++ gfx/drivers/d3d9cg.c
@@ -48,6 +48,8 @@
 #endif
 
 #include <retro_inline.h>
+#include <retro_common_api.h>
+#include <boolean.h>
 
 #include <defines/d3d_defines.h>
 #include "../common/d3d_common.h"
@@ -79,7 +81,426 @@
 #endif
 
 #include "d3d_shaders/opaque.cg.d3d9.h"
-#include "d3d9_renderchain.h"
+
+#define D3D_FILTER_LINEAR           (3 << 0)
+#define D3D_FILTER_POINT            (2 << 0)
+
+struct lut_info
+{
+   LPDIRECT3DTEXTURE9 tex;
+   char id[64];
+   bool smooth;
+};
+
+struct shader_pass
+{
+   unsigned last_width, last_height;
+   struct LinkInfo info;
+   D3DPOOL pool;
+   LPDIRECT3DTEXTURE9 tex;
+   LPDIRECT3DVERTEXBUFFER9 vertex_buf;
+   LPDIRECT3DVERTEXDECLARATION9 vertex_decl;
+   void *attrib_map;
+   void *vprg;
+   void *fprg;
+   void *vtable;
+   void *ftable;
+};
+
+#define D3D_PI 3.14159265358979323846264338327
+
+#define VECTOR_LIST_TYPE unsigned
+#define VECTOR_LIST_NAME unsigned
+#include "../../libretro-common/lists/vector_list.c"
+#undef VECTOR_LIST_TYPE
+#undef VECTOR_LIST_NAME
+
+#define VECTOR_LIST_TYPE struct lut_info
+#define VECTOR_LIST_NAME lut_info
+#include "../../libretro-common/lists/vector_list.c"
+#undef VECTOR_LIST_TYPE
+#undef VECTOR_LIST_NAME
+
+#define VECTOR_LIST_TYPE struct shader_pass
+#define VECTOR_LIST_NAME shader_pass
+#include "../../libretro-common/lists/vector_list.c"
+#undef VECTOR_LIST_TYPE
+#undef VECTOR_LIST_NAME
+
+typedef struct d3d9_renderchain
+{
+   unsigned pixel_size;
+   uint64_t frame_count;
+   struct
+   {
+      LPDIRECT3DTEXTURE9 tex[TEXTURES];
+      LPDIRECT3DVERTEXBUFFER9 vertex_buf[TEXTURES];
+      unsigned ptr;
+      unsigned last_width[TEXTURES];
+      unsigned last_height[TEXTURES];
+   } prev;
+   LPDIRECT3DDEVICE9 dev;
+   D3DVIEWPORT9 *out_vp;
+   struct shader_pass_vector_list  *passes;
+   struct unsigned_vector_list *bound_tex;
+   struct unsigned_vector_list *bound_vert;
+   struct lut_info_vector_list *luts;
+} d3d9_renderchain_t;
+
+static void *d3d9_texture_new(void *_dev,
+      unsigned width, unsigned height,
+      unsigned miplevels, unsigned usage, INT32 format,
+      INT32 pool, unsigned filter, unsigned mipfilter,
+      INT32 color_key, void *src_info_data,
+      PALETTEENTRY *palette, bool want_mipmap)
+{
+   LPDIRECT3DDEVICE9 dev = (LPDIRECT3DDEVICE9)_dev;
+   void *buf             = NULL;
+   if (want_mipmap)
+      usage |= D3DUSAGE_AUTOGENMIPMAP;
+   return (SUCCEEDED(IDirect3DDevice9_CreateTexture(dev,
+               width, height, miplevels, usage,
+               (D3DFORMAT)format,
+               (D3DPOOL)pool,
+               (struct IDirect3DTexture9**)&buf, NULL))) ? buf : NULL;
+}
+
+static INLINE bool d3d9_renderchain_add_pass(d3d9_renderchain_t *chain,
+      struct shader_pass *pass,
+      const struct LinkInfo *info)
+{
+   LPDIRECT3DTEXTURE9      tex;
+   LPDIRECT3DVERTEXBUFFER9 vertbuf = NULL;
+
+   if (!SUCCEEDED(IDirect3DDevice9_CreateVertexBuffer(
+               chain->dev, 4 * sizeof(struct Vertex),
+               D3DUSAGE_WRITEONLY, 0,
+               D3DPOOL_DEFAULT,
+               (LPDIRECT3DVERTEXBUFFER9*)&vertbuf, NULL)))
+      vertbuf = NULL;
+
+   if (!vertbuf)
+      return false;
+
+   pass->vertex_buf = vertbuf;
+
+   tex = (LPDIRECT3DTEXTURE9)d3d9_texture_new(
+         chain->dev,
+         info->tex_w,
+         info->tex_h,
+         1,
+         D3DUSAGE_RENDERTARGET,
+         (chain->passes->data[
+         chain->passes->count - 1].info.pass->fbo.flags & FBO_SCALE_FLAG_FP_FBO)
+         ? D3DFMT_A32B32G32R32F : D3D9_ARGB8888_FORMAT,
+         D3DPOOL_DEFAULT, 0, 0, 0, NULL, NULL, false);
+
+   if (!tex)
+      return false;
+
+   pass->tex        = tex;
+
+   IDirect3DDevice9_SetTexture(chain->dev, 0, (IDirect3DBaseTexture9*)pass->tex);
+   IDirect3DDevice9_SetSamplerState(chain->dev, 0, D3DSAMP_ADDRESSU, D3DTADDRESS_BORDER);
+   IDirect3DDevice9_SetSamplerState(chain->dev, 0, D3DSAMP_ADDRESSV, D3DTADDRESS_BORDER);
+   IDirect3DDevice9_SetTexture(chain->dev, 0, (IDirect3DBaseTexture9*)NULL);
+
+   shader_pass_vector_list_append(chain->passes, *pass);
+
+   return true;
+}
+
+static INLINE bool d3d9_renderchain_add_lut(d3d9_renderchain_t *chain,
+      const char *id, const char *path, bool smooth)
+{
+   struct lut_info info;
+   struct texture_image image;
+   LPDIRECT3DTEXTURE9 lut    = NULL;
+
+   image.pixels              = NULL;
+   image.width               = 0;
+   image.height              = 0;
+   image.supports_rgba       = true;
+
+   if (!image_texture_load(&image, path))
+   {
+      RARCH_ERR("[D3D9] Failed to load LUT image: %s.\n", path);
+      return false;
+   }
+
+   lut = (LPDIRECT3DTEXTURE9)d3d9_texture_new(
+         chain->dev,
+         image.width, image.height, 1,
+         0, D3D9_ARGB8888_FORMAT,
+         D3DPOOL_MANAGED, 0, 0, 0,
+         NULL, NULL, false);
+
+   if (!lut)
+   {
+      image_texture_free(&image);
+      return false;
+   }
+
+   {
+      D3DLOCKED_RECT lr;
+      if (SUCCEEDED(IDirect3DTexture9_LockRect(lut, 0, &lr, NULL, 0)))
+      {
+         unsigned y;
+         uint32_t       *dst   = (uint32_t*)lr.pBits;
+         const uint32_t *src   = image.pixels;
+         unsigned        pitch = lr.Pitch >> 2;
+         for (y = 0; y < image.height; y++, dst += pitch, src += image.width)
+            memcpy(dst, src, image.width << 2);
+         IDirect3DTexture9_UnlockRect(lut, 0);
+      }
+   }
+
+   image_texture_free(&image);
+
+   RARCH_LOG("[D3D9] LUT texture loaded: %s.\n", path);
+
+   info.tex    = lut;
+   info.smooth = smooth;
+   strlcpy(info.id, id, sizeof(info.id));
+
+   IDirect3DDevice9_SetTexture(chain->dev, 0, (IDirect3DBaseTexture9*)lut);
+   IDirect3DDevice9_SetSamplerState(chain->dev, 0, D3DSAMP_ADDRESSU, D3DTADDRESS_BORDER);
+   IDirect3DDevice9_SetSamplerState(chain->dev, 0, D3DSAMP_ADDRESSV, D3DTADDRESS_BORDER);
+   IDirect3DDevice9_SetTexture(chain->dev, 0, (IDirect3DBaseTexture9*)NULL);
+
+   lut_info_vector_list_append(chain->luts, info);
+
+   return true;
+}
+
+static INLINE void d3d9_renderchain_destroy_passes_and_luts(
+      d3d9_renderchain_t *chain)
+{
+   if (chain->passes)
+   {
+      int i;
+
+      for (i = 0; i < (int) chain->passes->count; i++)
+      {
+         if (chain->passes->data[i].attrib_map)
+            free(chain->passes->data[i].attrib_map);
+      }
+
+      shader_pass_vector_list_free(chain->passes);
+
+      chain->passes = NULL;
+   }
+
+   lut_info_vector_list_free(chain->luts);
+   unsigned_vector_list_free(chain->bound_tex);
+   unsigned_vector_list_free(chain->bound_vert);
+
+   chain->luts       = NULL;
+   chain->bound_tex  = NULL;
+   chain->bound_vert = NULL;
+}
+
+static INLINE void d3d9_renderchain_add_lut_internal(
+      d3d9_renderchain_t *chain,
+      unsigned index, unsigned i)
+{
+   /* 2 = D3D_TEXTURE_FILTER_LINEAR, 1 = D3D_TEXTURE_FILTER_POINT */
+   int32_t filter = chain->luts->data[i].smooth ? 2 : 1;
+   IDirect3DDevice9_SetTexture(chain->dev, index, (IDirect3DBaseTexture9*)chain->luts->data[i].tex);
+   IDirect3DDevice9_SetSamplerState(chain->dev, index, D3DSAMP_MAGFILTER, filter);
+   IDirect3DDevice9_SetSamplerState(chain->dev, index, D3DSAMP_MINFILTER, filter);
+   IDirect3DDevice9_SetSamplerState(chain->dev, index, D3DSAMP_ADDRESSU, D3DTADDRESS_BORDER);
+   IDirect3DDevice9_SetSamplerState(chain->dev, index, D3DSAMP_ADDRESSV, D3DTADDRESS_BORDER);
+   unsigned_vector_list_append(chain->bound_tex, index);
+}
+
+static INLINE void d3d9_renderchain_start_render(d3d9_renderchain_t *chain)
+{
+   chain->passes->data[0].tex         = chain->prev.tex[
+      chain->prev.ptr];
+   chain->passes->data[0].vertex_buf  = chain->prev.vertex_buf[
+      chain->prev.ptr];
+   chain->passes->data[0].last_width  = chain->prev.last_width[
+      chain->prev.ptr];
+   chain->passes->data[0].last_height = chain->prev.last_height[
+      chain->prev.ptr];
+}
+
+static INLINE void d3d9_renderchain_end_render(d3d9_renderchain_t *chain)
+{
+   chain->prev.last_width[chain->prev.ptr]  = chain->passes->data[0].last_width;
+   chain->prev.last_height[chain->prev.ptr] = chain->passes->data[0].last_height;
+   chain->prev.ptr                          = (chain->prev.ptr + 1) & TEXTURESMASK;
+}
+
+static INLINE void d3d9_renderchain_unbind_all(d3d9_renderchain_t *chain)
+{
+   int i;
+   /* Have to be a bit anal about it.
+    * Render targets hate it when they have filters apparently.
+    */
+   for (i = 0; i < (int) chain->bound_tex->count; i++)
+   {
+      IDirect3DDevice9_SetSamplerState(chain->dev,
+            chain->bound_tex->data[i], D3DSAMP_MINFILTER, D3DTEXF_POINT);
+      IDirect3DDevice9_SetSamplerState(chain->dev,
+            chain->bound_tex->data[i], D3DSAMP_MAGFILTER, D3DTEXF_POINT);
+      IDirect3DDevice9_SetTexture(chain->dev,
+            chain->bound_tex->data[i], (IDirect3DBaseTexture9*)NULL);
+   }
+
+   for (i = 0; i < (int) chain->bound_vert->count; i++)
+      IDirect3DDevice9_SetStreamSource(chain->dev, chain->bound_vert->data[i], 0, 0, 0);
+
+   if (chain->bound_tex)
+   {
+      unsigned_vector_list_free(chain->bound_tex);
+      chain->bound_tex = unsigned_vector_list_new();
+   }
+
+   if (chain->bound_vert)
+   {
+      unsigned_vector_list_free(chain->bound_vert);
+      chain->bound_vert = unsigned_vector_list_new();
+   }
+}
+
+static INLINE bool d3d9_renderchain_set_pass_size(
+      LPDIRECT3DDEVICE9 dev,
+      struct shader_pass *pass,
+      struct shader_pass *pass2,
+      unsigned width, unsigned height)
+{
+   if (width != pass->info.tex_w || height != pass->info.tex_h)
+   {
+      IDirect3DTexture9_Release(pass->tex);
+
+      pass->info.tex_w = width;
+      pass->info.tex_h = height;
+      pass->pool       = D3DPOOL_DEFAULT;
+      pass->tex        = (LPDIRECT3DTEXTURE9)
+         d3d9_texture_new(dev,
+            width, height, 1,
+            D3DUSAGE_RENDERTARGET,
+            (pass2->info.pass->fbo.flags & FBO_SCALE_FLAG_FP_FBO)
+            ? D3DFMT_A32B32G32R32F
+            : D3D9_ARGB8888_FORMAT,
+            D3DPOOL_DEFAULT, 0, 0, 0,
+            NULL, NULL, false);
+
+      if (!pass->tex)
+         return false;
+
+      IDirect3DDevice9_SetTexture(dev, 0, (IDirect3DBaseTexture9*)pass->tex);
+      IDirect3DDevice9_SetSamplerState(dev, 0, D3DSAMP_ADDRESSU, D3DTADDRESS_BORDER);
+      IDirect3DDevice9_SetSamplerState(dev, 0, D3DSAMP_ADDRESSV, D3DTADDRESS_BORDER);
+      IDirect3DDevice9_SetTexture(dev, 0, (IDirect3DBaseTexture9*)NULL);
+   }
+
+   return true;
+}
+
+static INLINE void d3d9_convert_geometry(
+      const struct LinkInfo *info,
+      unsigned *out_width,
+      unsigned *out_height,
+      unsigned width,
+      unsigned height,
+      D3DVIEWPORT9 *out_vp)
+{
+   switch (info->pass->fbo.type_x)
+   {
+      case RARCH_SCALE_VIEWPORT:
+         *out_width = info->pass->fbo.scale_x * out_vp->Width;
+         break;
+
+      case RARCH_SCALE_ABSOLUTE:
+         *out_width = info->pass->fbo.abs_x;
+         break;
+
+      case RARCH_SCALE_INPUT:
+         *out_width = info->pass->fbo.scale_x * width;
+         break;
+   }
+
+   switch (info->pass->fbo.type_y)
+   {
+      case RARCH_SCALE_VIEWPORT:
+         *out_height = info->pass->fbo.scale_y * out_vp->Height;
+         break;
+
+      case RARCH_SCALE_ABSOLUTE:
+         *out_height = info->pass->fbo.abs_y;
+         break;
+
+      case RARCH_SCALE_INPUT:
+         *out_height = info->pass->fbo.scale_y * height;
+         break;
+   }
+}
+
+static INLINE void d3d9_recompute_pass_sizes(
+      LPDIRECT3DDEVICE9 dev,
+      d3d9_renderchain_t *chain,
+      d3d9_video_t *d3d)
+{
+   int i;
+   struct LinkInfo link_info;
+   unsigned input_scale              = d3d->video_info.input_scale
+      * RARCH_SCALE_BASE;
+   unsigned current_width            = input_scale;
+   unsigned current_height           = input_scale;
+   unsigned out_width                = 0;
+   unsigned out_height               = 0;
+
+   link_info.pass                    = &d3d->shader.pass[0];
+   link_info.tex_w                   = current_width;
+   link_info.tex_h                   = current_height;
+
+   if (!d3d9_renderchain_set_pass_size(dev,
+            (struct shader_pass*)&chain->passes->data[0],
+            (struct shader_pass*)&chain->passes->data[
+            chain->passes->count - 1],
+            current_width, current_height))
+   {
+      RARCH_ERR("[D3D9] Failed to set pass size.\n");
+      return;
+   }
+
+   for (i = 1; i < (int) d3d->shader.passes; i++)
+   {
+      d3d9_convert_geometry(
+            &link_info,
+            &out_width, &out_height,
+            current_width, current_height, &d3d->out_vp);
+
+      link_info.tex_w = next_pow2(out_width);
+      link_info.tex_h = next_pow2(out_height);
+
+      if (!d3d9_renderchain_set_pass_size(dev,
+               (struct shader_pass*)&chain->passes->data[i],
+               (struct shader_pass*)&chain->passes->data[
+               chain->passes->count - 1],
+               link_info.tex_w, link_info.tex_h))
+      {
+         RARCH_ERR("[D3D9] Failed to set pass size.\n");
+         return;
+      }
+
+      current_width  = out_width;
+      current_height = out_height;
+
+      link_info.pass = &d3d->shader.pass[i];
+   }
+}
+
+static INLINE void d3d9_init_renderchain(d3d9_renderchain_t *chain)
+{
+   chain->passes     = shader_pass_vector_list_new();
+   chain->luts       = lut_info_vector_list_new();
+   chain->bound_tex  = unsigned_vector_list_new();
+   chain->bound_vert = unsigned_vector_list_new();
+}
 
 #define DECL_FVF_COLOR(stream, offset, index) \
    { (WORD)(stream), (WORD)(offset * sizeof(float)), D3DDECLTYPE_FLOAT4, D3DDECLMETHOD_DEFAULT, \
@@ -114,16 +535,27 @@ static LPDIRECT3DVERTEXBUFFER9 d3d9_cg_menu_pipeline_vbo = NULL;
 
 static LPDIRECT3DTEXTURE9 d3d9_cg_white_texture = NULL;
 
-/* Bind the stock Cg vertex and pixel shaders.
- * This is the CG equivalent of d3d9_hlsl_bind_program. */
-static INLINE void d3d9_cg_bind_program(cg_renderchain_t *chain)
+
+static void d3d9_vertex_buffer_free(void *vertex_data, void *vertex_declaration)
 {
-   if (chain)
+   if (vertex_data)
    {
-      cgD3D9BindProgram((CGprogram)chain->stock_shader.fprg);
-      cgD3D9BindProgram((CGprogram)chain->stock_shader.vprg);
+      LPDIRECT3DVERTEXBUFFER9 buf =
+         (LPDIRECT3DVERTEXBUFFER9)vertex_data;
+      IDirect3DVertexBuffer9_Release(buf);
+      buf = NULL;
+   }
+
+   if (vertex_declaration)
+   {
+      LPDIRECT3DVERTEXDECLARATION9 vertex_decl =
+         (LPDIRECT3DVERTEXDECLARATION9)vertex_declaration;
+      IDirect3DVertexDeclaration9_Release(vertex_decl);
+      vertex_decl = NULL;
    }
 }
+
+
 
 /* Set the modelViewProj matrix on the stock Cg vertex shader
  * via its named parameter. The Cg runtime assigns hardware
@@ -373,7 +805,8 @@ static void gfx_display_d3d9_cg_draw(gfx_display_ctx_draw_t *draw,
 
          if (_chain)
          {
-            d3d9_cg_bind_program(_chain);
+            cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+            cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
 
             IDirect3DDevice9_DrawPrimitive(dev, D3DPT_COMM_TRIANGLESTRIP,
                   0, draw->coords->vertices - 2);
@@ -598,14 +1031,16 @@ static void gfx_display_d3d9_cg_draw_pipeline(gfx_display_ctx_draw_t *draw,
             Vertex *verts    = NULL;
             unsigned vcount  = ca->coords.vertices;
 
-            d3d9_cg_menu_pipeline_vbo = (LPDIRECT3DVERTEXBUFFER9)
-                  d3d9_vertex_buffer_new(
-                  d3d->dev,
-                  vcount * sizeof(Vertex),
-                  D3DUSAGE_WRITEONLY,
-                  0,
-                  D3DPOOL_DEFAULT,
-                  NULL);
+            {
+               void *_vbuf = NULL;
+               if (SUCCEEDED(IDirect3DDevice9_CreateVertexBuffer(
+                           d3d->dev,
+                           vcount * sizeof(Vertex),
+                           D3DUSAGE_WRITEONLY, 0,
+                           D3DPOOL_DEFAULT,
+                           (LPDIRECT3DVERTEXBUFFER9*)&_vbuf, NULL)))
+                  d3d9_cg_menu_pipeline_vbo = (LPDIRECT3DVERTEXBUFFER9)_vbuf;
+            }
 
             if (d3d9_cg_menu_pipeline_vbo)
             {
@@ -1038,9 +1473,7 @@ static void d3d9_cg_font_render_msg(
    unsigned          width  = 0;
    unsigned          height = 0;
 
-   if (!font || !msg || !*msg)
-      return;
-   if (!d3d)
+   if (!d3d || !font || !msg || !*msg)
       return;
 
    video_driver_get_size(&width, &height);
@@ -1159,7 +1592,11 @@ static void d3d9_cg_font_render_msg(
          (LPDIRECT3DVERTEXDECLARATION9)d3d->menu_display.decl);
    {
       cg_renderchain_t *_chain = (cg_renderchain_t*)d3d->renderchain_data;
-      d3d9_cg_bind_program(_chain);
+      if (_chain)
+      {
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
+      }
    }
 
    {
@@ -1584,8 +2021,9 @@ static bool d3d9_cg_renderchain_init_shader_fvf(
 
    free(indices);
 
-   return d3d9_vertex_declaration_new(chain->dev,
-         decl, (void**)&pass->vertex_decl);
+   return (SUCCEEDED(IDirect3DDevice9_CreateVertexDeclaration(chain->dev,
+               (const D3DVERTEXELEMENT9*)decl,
+               (IDirect3DVertexDeclaration9**)&pass->vertex_decl)));
 }
 
 static void d3d9_cg_renderchain_bind_orig(
@@ -1943,10 +2381,16 @@ static bool d3d9_cg_renderchain_create_first_pass(
    {
       chain->prev.last_width[i]  = 0;
       chain->prev.last_height[i] = 0;
-      chain->prev.vertex_buf[i]  = (LPDIRECT3DVERTEXBUFFER9)
-         d3d9_vertex_buffer_new(
-            chain->dev, 4 * sizeof(struct D3D9CGVertex),
-            D3DUSAGE_WRITEONLY, 0, D3DPOOL_DEFAULT, NULL);
+      chain->prev.vertex_buf[i]  = NULL;
+         {
+            void *_vbuf = NULL;
+            if (SUCCEEDED(IDirect3DDevice9_CreateVertexBuffer(
+                        chain->dev, 4 * sizeof(struct D3D9CGVertex),
+                        D3DUSAGE_WRITEONLY, 0,
+                        D3DPOOL_DEFAULT,
+                        (LPDIRECT3DVERTEXBUFFER9*)&_vbuf, NULL)))
+               chain->prev.vertex_buf[i] = (LPDIRECT3DVERTEXBUFFER9)_vbuf;
+         }
 
       if (!chain->prev.vertex_buf[i])
          return false;
@@ -2371,7 +2815,11 @@ static void d3d9_cg_renderchain_render(
       IDirect3DSurface9_Release(back_buffer);
 
    d3d9_renderchain_end_render(chain);
-   d3d9_cg_bind_program(_chain);
+   if (_chain)
+   {
+      cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+      cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
+   }
    d3d9_cg_renderchain_calc_and_set_shader_mvp(
          (CGprogram)_chain->stock_shader.vprg,
          chain->out_vp->Width,
@@ -2839,19 +3287,26 @@ static bool d3d9_cg_initialize(d3d9_video_t *d3d, const video_info_t *info)
             D3DDECLUSAGE_COLOR, 0},
          D3DDECL_END()
       };
-      if (!d3d9_vertex_declaration_new(d3d->dev,
-               (void*)VertexElements, (void**)&d3d->menu_display.decl))
+      if (!SUCCEEDED(IDirect3DDevice9_CreateVertexDeclaration(d3d->dev,
+                  (const D3DVERTEXELEMENT9*)VertexElements,
+                  (IDirect3DVertexDeclaration9**)&d3d->menu_display.decl)))
          return false;
    }
 
    d3d->menu_display.offset = 0;
    d3d->menu_display.size   = 1024;
-   d3d->menu_display.buffer = d3d9_vertex_buffer_new(
-         d3d->dev, d3d->menu_display.size * sizeof(Vertex),
-         D3DUSAGE_WRITEONLY,
-         D3DFVF_XYZRHW | D3DFVF_DIFFUSE | D3DFVF_TEX1,
-         D3DPOOL_DEFAULT,
-         NULL);
+   {
+      void *_vbuf = NULL;
+      if (SUCCEEDED(IDirect3DDevice9_CreateVertexBuffer(
+                  d3d->dev, d3d->menu_display.size * sizeof(Vertex),
+                  D3DUSAGE_WRITEONLY,
+                  D3DFVF_XYZRHW | D3DFVF_DIFFUSE | D3DFVF_TEX1,
+                  D3DPOOL_DEFAULT,
+                  (LPDIRECT3DVERTEXBUFFER9*)&_vbuf, NULL)))
+         d3d->menu_display.buffer = _vbuf;
+      else
+         d3d->menu_display.buffer = NULL;
+   }
 
    if (!d3d->menu_display.buffer)
       return false;
@@ -3136,7 +3591,7 @@ static void d3d9_cg_free_overlays(d3d9_video_t *d3d)
    d3d->overlays_size = 0;
 }
 
-static void d3d9_overlay_tex_geom(
+static void d3d9_cg_overlay_tex_geom(
       void *data,
       unsigned index,
       float x, float y,
@@ -3153,7 +3608,7 @@ static void d3d9_overlay_tex_geom(
    d3d->overlays[index].tex_coords[3] = h;
 }
 
-static void d3d9_overlay_vertex_geom(
+static void d3d9_cg_overlay_vertex_geom(
       void *data,
       unsigned index,
       float x, float y,
@@ -3172,7 +3627,7 @@ static void d3d9_overlay_vertex_geom(
    d3d->overlays[index].vert_coords[3] = h;
 }
 
-static bool d3d9_overlay_load(void *data,
+static bool d3d9_cg_overlay_load(void *data,
       const void *image_data, unsigned num_images)
 {
    unsigned i, y;
@@ -3221,14 +3676,14 @@ static bool d3d9_overlay_load(void *data,
       overlay->tex_h         = height;
 
       /* Default. Stretch to whole screen. */
-      d3d9_overlay_tex_geom(d3d, i, 0, 0, 1, 1);
-      d3d9_overlay_vertex_geom(d3d, i, 0, 0, 1, 1);
+      d3d9_cg_overlay_tex_geom(d3d, i, 0, 0, 1, 1);
+      d3d9_cg_overlay_vertex_geom(d3d, i, 0, 0, 1, 1);
    }
 
    return true;
 }
 
-static void d3d9_overlay_enable(void *data, bool state)
+static void d3d9_cg_overlay_enable(void *data, bool state)
 {
    unsigned i;
    d3d9_video_t            *d3d = (d3d9_video_t*)data;
@@ -3244,7 +3699,7 @@ static void d3d9_overlay_enable(void *data, bool state)
 #endif
 }
 
-static void d3d9_overlay_full_screen(void *data, bool enable)
+static void d3d9_cg_overlay_full_screen(void *data, bool enable)
 {
    unsigned i;
    d3d9_video_t *d3d = (d3d9_video_t*)data;
@@ -3253,26 +3708,26 @@ static void d3d9_overlay_full_screen(void *data, bool enable)
       d3d->overlays[i].fullscreen = enable;
 }
 
-static void d3d9_overlay_set_alpha(void *data, unsigned index, float mod)
+static void d3d9_cg_overlay_set_alpha(void *data, unsigned index, float mod)
 {
    d3d9_video_t *d3d = (d3d9_video_t*)data;
    if (d3d)
       d3d->overlays[index].alpha_mod = mod;
 }
 
-static const video_overlay_interface_t d3d9_overlay_interface = {
-   d3d9_overlay_enable,
-   d3d9_overlay_load,
-   d3d9_overlay_tex_geom,
-   d3d9_overlay_vertex_geom,
-   d3d9_overlay_full_screen,
-   d3d9_overlay_set_alpha,
+static const video_overlay_interface_t d3d9_cg_overlay_interface = {
+   d3d9_cg_overlay_enable,
+   d3d9_cg_overlay_load,
+   d3d9_cg_overlay_tex_geom,
+   d3d9_cg_overlay_vertex_geom,
+   d3d9_cg_overlay_full_screen,
+   d3d9_cg_overlay_set_alpha,
 };
 
 void d3d9_cg_get_overlay_interface(void *data,
       const video_overlay_interface_t **iface)
 {
-   *iface = &d3d9_overlay_interface;
+   *iface = &d3d9_cg_overlay_interface;
 }
 #endif
 
@@ -3328,7 +3783,7 @@ static void d3d9_cg_free(void *data)
 
 /* Standalone overlay render for the D3D9 Cg driver.
  *
- * The common d3d9_overlay_render() sets its MVP via
+ * The common d3d9_cg_overlay_render() sets its MVP via
  * SetVertexShaderConstantF which does not work with Cg shaders
  * (the Cg runtime manages its own constant register mapping).
  * This reimplementation uses the Cg API to set the MVP correctly
@@ -3370,7 +3825,11 @@ static void d3d9_cg_overlay_render(
           0.0f,  0.0f, 0.0f,  1.0f
       };
       cg_renderchain_t *_chain = (cg_renderchain_t*)d3d->renderchain_data;
-      d3d9_cg_bind_program(_chain);
+      if (_chain)
+      {
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
+      }
       d3d9_cg_set_mvp(_chain, bottomup_ortho);
    }
 
@@ -3551,7 +4010,11 @@ static bool d3d9_cg_frame(void *data, const void *frame,
       IDirect3DDevice9_SetViewport(d3d->dev, (D3DVIEWPORT9*)&screen_vp);
       {
          cg_renderchain_t *_chain = (cg_renderchain_t*)d3d->renderchain_data;
-         d3d9_cg_bind_program(_chain);
+         if (_chain)
+         {
+            cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+            cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
+         }
       }
       IDirect3DDevice9_BeginScene(d3d->dev);
       menu_driver_frame(menu_is_alive, video_info);
@@ -3610,7 +4073,10 @@ static bool d3d9_cg_frame(void *data, const void *frame,
       IDirect3DDevice9_SetRenderState(d3d->dev,
             D3DRS_CULLMODE, D3DCULL_NONE);
       if (_chain)
-         d3d9_cg_bind_program(_chain);
+      {
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.fprg);
+         cgD3D9BindProgram((CGprogram)_chain->stock_shader.vprg);
+      }
       IDirect3DDevice9_BeginScene(d3d->dev);
       gfx_widgets_frame(video_info);
       IDirect3DDevice9_EndScene(d3d->dev);
@@ -3745,10 +4211,8 @@ static void d3d9_cg_set_video_mode(void *data,
       unsigned width, unsigned height,
       bool fullscreen)
 {
-#ifndef _XBOX
    /* TODO/FIXME - why is this called here? */
    win32_show_cursor(data, !fullscreen);
-#endif
 }
 
 static const video_poke_interface_t d3d9_cg_poke_interface = {
@@ -3885,13 +4349,9 @@ static INLINE bool d3d9_cg_device_get_render_target_data(
       LPDIRECT3DDEVICE9 dev,
       LPDIRECT3DSURFACE9 src, LPDIRECT3DSURFACE9 dst)
 {
-#ifndef _XBOX
    return (   dev
            && SUCCEEDED(IDirect3DDevice9_GetRenderTargetData(
                dev, src, dst)));
-#else
-   return false;
-#endif
 }
 
 bool d3d9_cg_read_viewport(void *data, uint8_t *buffer, bool is_idle)
@@ -3954,11 +4414,7 @@ end:
 
 static bool d3d9_cg_has_windowed(void *data)
 {
-#ifdef _XBOX
-   return false;
-#else
    return true;
-#endif
 }
 
 video_driver_t video_d3d9_cg = {
